requirements.txt
python-dotenv
pandas
numpy
scikit-learn
scipy
# pymongo[srv]==3.6
pymongo[srv]
pymongo
typing-extensions
certifi
setuptools
pyyaml
pathlib
dill
mlflow
fastapi
uvicorn
python-multipart
# -e .